requirements.txt
appdirs==1.4.4
asgiref==3.2.3
autopep8==1.5.7
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cssselect==1.1.0
cssutils==2.2.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-cookies-samesite==0.9.0
django-cors-headers==3.5.0
django-debug-toolbar==3.2.1
django-heroku==0.3.1
django-samesite-none==0.0.3
djangorestframework==3.11.0
django_rest_knox==4.1.0
Faker==8.9.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
MouseInfo==0.1.3
mysqlclient==1.4.6
numpy==1.19.4
Pillow==8.0.1
premailer==3.7.0
psycopg2==2.8.6
PyAutoGUI==0.9.52
pycodestyle==2.7.0
PyGetWindow==0.0.9
PyJWT==2.1.0
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-for-android==0.4
PyTweening==1.0.3
pytz==2019.3
pywhatkit==3.9
requests==2.25.1
sh==1.14.1
six==1.15.0
soupsieve==2.1
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.2
ua-parser==0.10.0
urllib3==1.26.2
virtualenv==20.2.2
whitenoise==5.2.0
wikipedia==1.4.0
yagmail==0.14.245
zipp==3.4.0
